+++ Makefile.local
@@ -1,6 +1,6 @@
 CMAKE_COMMON_FLAGS += -DUSERVER_FEATURE_CRYPTOPP_BLAKE=0 -DUSERVER_DOWNLOAD_PACKAGE_CRYPTOPP=1
 # Enable grpc
-CMAKE_COMMON_FLAGS += -DUSERVER_FEATURE_GRPC=1 -DUSERVER_FEATURE_GRPC_CHANNELZ=0
+CMAKE_COMMON_FLAGS += -DUSERVER_FEATURE_GRPC=1
 # On for Arch
 # CMAKE_COMMON_FLAGS += -DUSERVER_FEATURE_PATCH_LIBPQ=0
 # Off not using features
