+++ src/hello_client.hpp
@@ -3,6 +3,7 @@
 #include <handlers/hello_client.usrv.pb.hpp>
 #include <userver/components/component_list.hpp>
 #include <userver/components/loggable_component_base.hpp>
+#include <userver/components/impl/component_base.hpp>
 #include <userver/ugrpc/client/client_factory_component.hpp>
 
 namespace pg_grpc_service_template {
@@ -33,4 +34,8 @@ class HelloClient final : public userver::components::LoggableComponentBase {
 
 void AppendHelloClient(userver::components::ComponentList& component_list);
 
+
 }  // namespace pg_grpc_service_template
+
+template <>
+inline constexpr bool userver::components::kHasValidate<pg_grpc_service_template::HelloClient> = true;
