+++ tests/test_grpc_server.py
@@ -1,9 +1,19 @@
 import pytest
 
-from testsuite.databases import pgsql
-
-
 # Start the tests via `make test-debug` or `make test-release`
+
+
+async def test_grpc_client(hello_protos, mock_grpc_hello, grpc_service):
+    @mock_grpc_hello('SayHello')
+    async def mock_say_hello(request, context):
+        return hello_protos.HelloResponse(
+            text=f'{request.name}!!',
+        )
+
+    request = hello_protos.HelloRequest(name='mock_userver')
+    response = await grpc_service.SayHello(request)
+    assert response.text == 'Hello, userver!!!\n'
+    assert mock_say_hello.times_called == 1
 
 
 async def test_first_time_users(hello_protos, grpc_service):
